# Go/GoServer/go/model_planet.go
/*
 * 4th Year Project
 *
 * (Will come up with a better name later)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Planet struct {

	Name string `json:"name"`

	RotationPeriod float64 `json:"rotation_period,omitempty"`

	OrbitalPeriod float64 `json:"orbital_period,omitempty"`

	Diameter int32 `json:"diameter,omitempty"`

	Climate string `json:"climate,omitempty"`

	Gravity string `json:"gravity,omitempty"`

	Terrain string `json:"terrain,omitempty"`

	SurfaceWater float64 `json:"surface_water,omitempty"`

	Population int32 `json:"population,omitempty"`
}